# [old] Overhaul sim/dex* initialization