… server/chat-plugins/othermetas.ts
… sim/dex-conditions.ts
… sim/dex-data.ts
… sim/dex-formats.ts
… sim/dex-items.ts
… sim/dex-moves.ts
… sim/dex-species.ts
… sim/dex.ts
… sim/tools/exhaustive-runner.ts
… test/common.js
… test/sim/data.js
… test/sim/team-validator/basic.js
… tools/set-import/importer.ts
… tools/simulate/index.js