# Repurposed
Lot of optimizations were explored here. They will be factored out into small PRs.

Huge thanks to slayer95 and urkerab for their feedback.

## Features
* Lower memory usage
* Better use of V8's 'hidden classes'
* More predictable loading order
* faster toID
* cached empty objects

## Bug fixes
`DexStats` now properly checks dex gen during its constructor.

## High level changes
The external API described in [DEX.md](https://github.com/smogon/pokemon-showdown/blob/master/sim/DEX.md) should be unaffected.

### Before
* Eagerly constructed mods, lazily constructed data
* Inheritance by recursive copying
* Unordered field initialization in ctors

### After
* Lazily constructed mods, eagerly constructed data
* Inheritance by Copy-on-Write
* Consistent field initialization when possible
  * Helps with 'hidden classes' in V8


## Migration guide
Applicable to code that depended on implementation details of Dex.

### `dex.loadData()`, `dex.includeFormat()`
No longer necessary, just use `dex.mod(modname)` when you need a mod.

### `dex.dexes`
No.  
Use `Dex.mod(_)` for retrieving mods and `Dex.scanMods()` to see all possible mods.