@larry-the-table-guy 2024-09-13
Results of above experiment on my machine:
Memory usage as mods are loaded:
| # mods | Mem |
| ------ | --- |
|  5     |  26 |
| 10     |  51 |
| 15     | 109 |
| 20     | 144 |
| 30     | 210 |
| 38     | 239 |


<details>
<summary>Console.log output from the profiling code</summary>

```
	mod sub data: species
num unique ids:  1427
key: #objs per id, value: how many items had this count Map(15) {
  7 => 285,
  9 => 151,
  15 => 7,
  10 => 130,
  5 => 127,
  8 => 188,
  6 => 257,
  12 => 73,
  11 => 120,
  13 => 47,
  14 => 19,
  4 => 16,
  3 => 1,
  1 => 5,
  2 => 1
}
amt added by each mod (depends on mod order): Map(38) {
  'base' => 1421,
  'gen9' => 0,
  'fullpotential' => 0,
  'gen1' => 1421,
  'gen1jpn' => 0,
  'gen1rbycap' => 5,
  'gen1stadium' => 68,
  'gen2' => 1420,
  'gen2stadium2' => 0,
  'gen3' => 429,
  'gen4' => 374,
  'gen4pt' => 0,
  'gen5' => 974,
  'gen5bw1' => 188,
  'gen6' => 502,
  'gen6megasrevisited' => 44,
  'gen6xy' => 17,
  'gen7' => 448,
  'gen7letsgo' => 1421,
  'gen7pokebilities' => 0,
  'gen7sm' => 11,
  'gen8' => 650,
  'gen8bdsp' => 346,
  'gen8dlc1' => 150,
  'gen8linked' => 0,
  'gen9dlc1' => 373,
  'gen9predlc' => 266,
  'gen9ssb' => 201,
  'gennext' => 59,
  'mixandmega' => 48,
  'partnersincrime' => 0,
  'pokebilities' => 0,
  'pokemoves' => 0,
  'randomroulette' => 0,
  'sharedpower' => 0,
  'sharingiscaring' => 0,
  'thecardgame' => 751,
  'trademarked' => 0
}
num unique objects: 52585
num unique value: 11587
```
</details>