@Slayer95 2024-09-14
There are indeed a lot of duplicate```Species``, because of the current [special-casing](https://github.com/smogon/pokemon-showdown/blob/master/sim/dex.ts#L527-L533) of ``Pokedex`` in the data load algorithm. It's nice that you are able to measure the magnitude of this issue.

Given the developments in the Dex data modules over the recent years (mainly, the ``Species`` constructor, meaning that ``data.Pokedex`` entries are no longer overriden by ``FormatsData``, and ``Learnsets``), this issue can probably be safely fixed now.

Shallow clones do rise the stakes for when inexperienced modders forget using ``modData()`` in their ``scripts.ts`` files, but that has always been an acceptable risk, imo.