@larry-the-table-guy 2024-09-14
@Slayer95 Thanks for looking into it.

I had tried removing that special case in the morning but forgot to document it. Got this result
```
num unique objects: 52585
num unique value: 12614
```
I see the git blame has that code at ~9 years old; clearly much of the design has changed since then, as you mentioned.

As for how to approach this, I think it helps to view the data as the ground truth.
Here is the size of `./data/` after `tar -cOI 'xz -9'`
```
1,503.0 KiB / 27.2 MiB = 0.054
```
Indicating that there is only 1.5 MB of real information. Certainly, we won't get that low, but even 20x overhead puts us at 30MB.

I'll have to spend some more time digging around `sim/`, but I would think this is more of a copy-on-write situation than memoization. AFAICT, we know exactly when the data changes.

Overall, I think the dex* files would really benefit from a bit of a database mentality - after all, their inputs are really just tables (no bias :P).