@Slayer95 2024-09-14
``num unique value: 11587 → 12614``

Unexpected, though not entirely unplausible. Something to investigate if we were to memoize. But since you plan to follow a different route, I guess it doesn't matter a lot.

> Overall, I think the dex* files would really benefit from a bit of a database mentality 

I can see what you mean. Following thi path, the gains can potentially be enormous.