@larry-the-table-guy 2024-10-08
Some simple, effective optimizations for `toID`.

1) Simpler type coercion that optimizes for the common case
```js
if (typeof text === 'string') {
        // 99.9% case, skip checks
} else {
        if (text) {
                text = text.id || text.userid || text.roomid || text;
        }
        if (typeof text === 'number') text = '' + text;
        else if (typeof text !== 'string') return '';
}
// text is now definitely a string
```

2) Fast path for strings that are already valid IDs
```js
// very often text is already a valid ID.
if (/^[a-z0-9]*$/.test(text)) {
        return text;
}
```

3) Cache common name->ID pairs
```js
// Next, we often produce the same IDs many times.
const c = _toIDCache.get(text);
if (c) {
        return c;
}
```

Lastly, do the generic case (unchanged).
```js
return text.toLowerCase().replace(/[^a-z0-9]+/g, '') as ID;
```

---

## Profiling results
TL;DR: really good.

time in ms. each row is stats for a batch of 4194304 calls to `toID`, while running `npm run full-test`. Already adjusted for overhead of timing.

|(0) original time | (1)  time | (2)  time | (2) alloc rate | (3) time | (3) alloc rate |
| ----------------- | ---------------- | ------- | -------- | ------- | -------- |
| 633               | 570              | 452     | 0.188    | 422     | 0.137    |
| 718               | 606              | 432     | 0.111    | 367     | 0.023    |
| 909               | 689              | 561     | 0.182    | 457     | 0.008    |
| 672               | 556              | 479     | 0.458    | 339     | 0.036    |
| 757               | 620              | 615     | 0.583    | 457     | 0.098    |
| 736               | 629              | 576     | 0.543    | 466     | 0.156    |
| 579               | 480              | 428     | 0.532    | 244     | 0        |
| 570               | 474              | 409     | 0.521    | 243     | 0        |
| 589               | 488              | 426     | 0.535    | 239     | 0        |
| 633               | 524              | 504     | 0.627    | 255     | 0        |
| 653               | 533              | 575     | 0.737    | 296     | 0.097    |
| 707               | 554              | 642     | 0.77     | 363     | 0.035    |
| 762               | 582              | 600     | 0.62     | 368     | 0        |
| 616               | 495              | 465     | 0.569    | 272     | 0        |
| 645               | 528              | 538     | 0.676    | 341     | 0.052    |
| 726               | 594              | 578     | 0.547    | 478     | 0.216    |
| 610               | 482              | 442     | 0.485    | 270     | 0        |
| 596               | 494              | 487     | 0.643    | 272     | 0        |
| 612               | 494              | 487     | 0.582    | 282     | 0        |
| 642               | 491              | 468     | 0.467    | 305     | 0        |
| 906               | 737              | 726     | 0.472    | 733     | 0.418    |
| 950               | 753              | 750     | 0.504    | 798     | 0.468    |
| 1007              | 761              | 743     | 0.503    | 788     | 0.466    |
| 905               | 770              | 739     | 0.498    | 799     | 0.463    |
| 925               | 763              | 745     | 0.498    | 796     | 0.466    |
| 937               | 777              | 751     | 0.494    | 824     | 0.459    |
| 947               | 767              | 763     | 0.495    | 825     | 0.466    |

Last ~30 million calls have a high alloc rate - if you dump the inputs the pattern you see is "${name} {number}", where name is typically in the id cache. So, if one wanted to improve that metric, you'd go to the call-sites and change how they call toID (you'd do something like `toID(name) + ' ' + number`, which should be cheaper than concatenating first and then hitting the generic case in toID.