@larry-the-table-guy 2024-10-12
Not reviving this, just did a comparison with the base commit (4731f1283) to see the impact of all the optimizations together.

Important note: this is **without** dedup for species. That would be another considerable bump.

## Time
| Action | Old time | New time |
| ------ | -------- | -------- |
| `npm test` | 16s | 10s |
| `npm run full-test` | 5m | 5m |
| (1/2) `Dex.includeModData` | 1.18s| - |
| (2/2) all dexes load | 2.69s | - |
| (1 & 2) merged loading | 3.87 | 2.50 |

## Memory
| Stage | Old | New |
| ----- | --- | --- |
| pre Dex | 7M | 10M |
| includeModData | 204M | - |
| + `.all()`s | 306M | 227M |


# Benchmark source code
`Dex.includeModData`
```ts
console.log(process.memoryUsage());
const start = performance.now();
Dex.includeMods();
Dex.includeModData();
console.log(performance.now() - start);
global.gc();
console.log(process.memoryUsage());
```

`all dexes load`
```ts
Dex.includeMods();
Dex.includeModData();
console.log(process.memoryUsage());
const start = performance.now();
const mods = ["base", "gen9", "fullpotential", "gen1", "gen1jpn", "gen1rbycap", "gen1stadium",
        "gen2", "gen2stadium2", "gen3", "gen4", "gen4pt", "gen5", "gen5bw1", "gen6", "gen6megasrevisited", "gen6xy",
        "gen7", "gen7letsgo", "gen7pokebilities", "gen7sm", "gen8", "gen8bdsp", "gen8dlc1", "gen8linked",
        "gen9dlc1", "gen9predlc", "gen9ssb", "gennext", "mixandmega", "partnersincrime", "pokebilities",
        "pokemoves", "randomroulette", "sharedpower", "sharingiscaring", "thecardgame", "trademarked",
];
for (const modname of mods) {
        const m = Dex.mod(modname);
        m.species.all();
        m.items.all();
        m.moves.all();
        m.abilities.all();
        m.natures.all();
        m.types.all();
}
console.log(performance.now() - start);
global.gc();
console.log(process.memoryUsage());
```

`merged loading`
put start counters before constructing base mod.
put end counters at the bottom of `dex.ts` (with previous logic for loading everything).